PROJ-48_ForgotRecoveryPasswordApi fix: move type to shared and update imports
