+++ shared/lib/handleRecoveryAndEmailLinkErrors/index.ts
@@ -1,0 +1,2 @@
+export * from './handleRecoveryAndEmailLinkErrors'
+export * from './types'
